requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
lxml>=5.1.0
python-docx>=1.1.0
docx2pdf>=0.1.8
pytest>=8.0.0
pytest-mock>=3.12.0
weasyprint